+++ ibl_alignment_gui/app/widgets/qrange_slider.py
