+++ ibl_alignment_gui/backends/allen/anatomical_atlas.py
@@ -2,6 +2,7 @@ import logging
 
 import numpy as np
 import SimpleITK as sitk
+
 from iblatlas.atlas import BrainAtlas, BrainCoordinates
 from iblatlas.regions import BrainRegions
 from iblutil.numerical import ismember
