+++ ibl_alignment_gui/loaders/histology_loader.py
@@ -9,8 +9,11 @@ import numpy as np
 import requests
 import SimpleITK as sitk  # noqa: N813
 
+from ibl_alignment_gui.backends.allen.anatomical_atlas import (
+    _BLESSED_DIRECTION,
+    BrainAtlasAnatomical,
+)
 from iblatlas.atlas import AllenAtlas, BrainAtlas
-from ibl_alignment_gui.utils.allen.anatomical_atlas import BrainAtlasAnatomical, _BLESSED_DIRECTION
 from iblutil.util import Bunch
 from one import params
 from one.webclient import http_download_file
