+++ ibl_alignment_gui/plugins/ephys_features.py
@@ -8,17 +8,17 @@ import pyqtgraph as pg
 from qtpy import QtCore, QtGui, QtWidgets
 
 from ibl_alignment_gui.app.shank_view import ShankView
-from ibl_alignment_gui.loaders.data_loader import FeatureLoaderOne
-from ibl_alignment_gui.loaders.geometry_loader import ChannelGeometry, arrange_channels_into_banks
-from ibl_alignment_gui.utils.ephys_alignment import EphysAlignment
-from ibl_alignment_gui.utils.qt.adapted_axis import replace_axis
-from ibl_alignment_gui.utils.qt.custom_widgets import (
+from ibl_alignment_gui.app.widgets.adapted_axis import replace_axis
+from ibl_alignment_gui.app.widgets.custom_widgets import (
     ColorBar,
     PopupWindow,
     SelectionWidget,
     SliderWidget,
     set_axis,
 )
+from ibl_alignment_gui.core.ephys_alignment import EphysAlignment
+from ibl_alignment_gui.loaders.data_loader import FeatureLoaderOne
+from ibl_alignment_gui.loaders.geometry_loader import ChannelGeometry, arrange_channels_into_banks
 from iblutil.util import Bunch
 
 if TYPE_CHECKING:
